setting up screens
